Move out the current InvocationHandler
